requirements.txt
uvicorn[standard]
fastapi
boto3>=1.26.0
sqlalchemy
psycopg[binary]
python-dotenv
alembic
pytest
httpx
pytest-asyncio
fastapi-users
celery
authlib
python-jose
asyncpg
itsdangerous
greenlet
